pyproject.toml
[tool.poetry]
name = "cot"
version = "0.1.0"
description = ""
authors = ["aginns <aginns@motorway.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
ruff = "^0.6.5"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
openai = "^1.45.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
